fix: pre-screening distribution bug
