+++ server/src/routes/course/stageInterview/index.ts
@@ -1,0 +1,11 @@
+export { createFeedback } from './createFeedback';
+export { createInterview } from './createInterview';
+export { createInterviews } from './createInterviews';
+export { createInterviewStudent } from './createInterviewStudent';
+export { deleteInterview } from './deleteInterview';
+export { getAvailableStudents } from './getAvailableStudents';
+export { getFeedback } from './getFeedback';
+export { getInterviewStudent } from './getInterviewStudent';
+export { getInterviewerStudents } from './getInterviewerStudents';
+export { getInterviews } from './getInterviews';
+export { updateInterview } from './updateInterview';
